Unify sandbox/remote handling of empty TreeArtifacts
Previously, SandboxHelpers contained special handling for the case of
empty TreeArtifact action inputs to ensure that they are added to the
sandbox as empty directories. As pointed out in a comment, this logic
should live in SpawnInputExpander, where it would also apply to remote
execution.

This commit adds an integration test for this previously untested case
and extracts the logic into SpawnInputExpander. It also updates some
comments that didn't mention TreeArtifact expansion even though it was
already performed before this commit.