Optionally assume that the SkyQuery universe is closed
With the new option `--assume_closed_universe`, SkyQuery assumes that
the repositories mentioned in the universe scope (explicit or inferred)
have no incoming or outgoing dependency edges to other repositories.
This can speed up the evaluation of queries by avoiding the need to
fetch repositories outside the universe, but can also lead to incorrect
results if the assumption is violated.