Convert server args to the internal string representation
The arguments for requests to the server were already converted to Bazel's internal string representation, which resulted in a mismatch between `--client_cwd` and `--workspace_directory` if the workspace path contains non-ASCII characters.

Since there are more places in the code base that need this conversion, this change introduces optimized uniform conversion functions to replace ad-hoc conversion logic.